+++ src/io/Reader.js
@@ -1,5 +1,5 @@
 import { Console } from '@woowacourse/mission-utils';
-import * as ERROR_MESSAGES from '../cosntants/errorMessages.js';
+import * as ERROR_MESSAGES from '../constants/errorMessages.js';
 import { throwError } from '../utils/errorHandler.js';
 
 class Reader {
